+++ features/FEATURE_BLE/ble/generic/GenericGap.h
@@ -116,8 +116,8 @@ public:
 
     virtual ble_error_t setPeriodicAdvertisingParameters(
         AdvHandle_t handle,
-        uint16_t periodicAdvertisingIntervalMinMs,
-        uint16_t periodicAdvertisingIntervalMaxMs,
+        uint32_t periodicAdvertisingIntervalMinMs,
+        uint32_t periodicAdvertisingIntervalMaxMs,
         bool advertiseTxPower
     );
